Changes to Project

+++ src/main/java/com/company/Crew.java
@@ -3,12 +3,11 @@ package com.company;
 public class Crew extends Person {
     private String department;
 
-    public Crew(String personName, int personSalary, String department) {
+    public Crew(String personName, double personSalary, String department) {
         super(personName, personSalary);
         this.department = department;
 
     }
-
     public String getDepartment() {
         return department;
     }
